+++ Makefile
@@ -1,4 +1,3 @@
-
 .PHONY: start-backing-services
 start-backing-services:
 	@echo "Starting backing services..."
@@ -21,3 +20,7 @@ test-hurl:
 sqlc-generate:
 	@echo "Generating SQLC code..."
 	sqlc generate
+
+.PHONY: test
+test:
+	go test -v ./...
